feat: Add Telescope picker for managing buf pin state
Adds a Telescope picker that displays all listed buffers and their pin status, indicated by a customizable icon. The following actions can be performed on any buffer(s) selected in the picker:

- `<M-y>`: Toggle the pin status of all selected buffers. Multiple buffers can be selected (default `<Tab>`). After the toggle action, the icon will be updated to reflect the new pin status. The picker does not close.
- `<M-a>`: Same as above, but pins/unpins all listed buffers
- `<M-x>`: Delete all selected buffers. The picker does not close. Uses the function defined in `opts.close_command`
- `<M-c>`: Close all unpinned buffers. The picker does not close.
- `<CR>`: Opens the selected buffer in the current window and closes the picker.

The picker depends on `telescope.nvim` and its dependency `plenary.nvim`. A `pcall` check is used to prevent the module from loading if these are not present.
The display style of the filepath in the picker is copied from [smart-open.nvim](https://github.com/danielfalk/smart-open.nvim) by @danielfalk.

The `opts` table adds configurable options for the picker in the `telescope` table:
- `mappings`: set keybindings for the actions described above
- `pin_icons`: set the pin/unpin icons and their respective highlight groups

Co-Authored-By: Aleksandr Kirillov <32141102+axkirillov@users.noreply.github.com>
Co-Authored-By: Tushar <hello@ditsuke.com>