feat(telescope): accept relevant buffers() opts for picker (#13)
This commit adds a new module, hbac.telescope.results_opts`, which filters the results of the entry picker to only show buffers that match the options passed to the picker. Options can be passed by passing a table as an argument to `hbac.telescope()`. This commit supports the options listed in `:h telescope.builtin.buffers()` and those in `:h telescope.setup()` relevant to the buffer picker. The notable exception is `mappings`, which will be handled in an upcoming commit, along with the ability to set default opts in the setup config table for the `Hbac telescope` command.

Example usage:

```lua
vim.keymap.set("n", ",ht", function()
    local default_opts = {
        sort_mru = true,
        sort_lastused = true,
    }
    require("hbac").telescope(default_opts)
end, { desc = "Hbac: pin picker", noremap = true })
```